requirements.txt
Flask
Flask-SQLAlchemy
alpaca-trade-api
Flask-Babel
python-dotenv
cryptography
requests
pytz
werkzeug